4.21 done

+++ controllers/blogs.js
@@ -43,9 +43,23 @@ blogsRouter.post('', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    const user = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'this blog does not belong to the currently logged in user' })
+    }
+    
   } catch (exception) {
     next(exception)
   }
